Clear clipboard selection from clippy history
This commit adds support to the timed clipboard clearing for the
`clippy` clipboard manager, which otherwise retains the password in
history, though the previous selection is restored to first place in
history.

Though, I wonder if a better approach might be to support a
`PASSWORD_STORE_CLIP_RM_CMD` environment variable (or similar) - or
even use a `clip_rm` function that could be overridden by an extension,
allowing somebody to write, for example, the `pass-clippy` extension
rather than every user figuring out the correct command independently
(or searching for something to copy).